Fix legit test failures caused in upgrade.js modules
We were accessing `bgblack` but it is `bgBlack`.

CI was previously green, which makes me think this change may have
occurred during a change in one of our transitive dependencies; here's a
snippet from the output of `yarn why ansi-colors` before this commit:

```
=> Found "ansi-colors@3.2.3"
info Has been hoisted to "ansi-colors"
info Reasons this module exists
   - Hoisted from "_project_#liferay-theme-es2015-hook#ansi-colors"
   - Hoisted from "_project_#liferay-theme-tasks#gulp-inject#ansi-colors"
=> Found "plugin-error#ansi-colors@1.1.0"
info This module exists because "_project_#liferay-theme-tasks#plugin-error" depends on it.
```

And after:

```
=> Found "ansi-colors@3.2.3"
info Reasons this module exists
   - "_project_#liferay-theme-es2015-hook" depends on it
   - Hoisted from "_project_#liferay-theme-es2015-hook#ansi-colors"
   - Hoisted from "_project_#liferay-theme-tasks#plugin-error#ansi-colors"
   - Hoisted from "_project_#liferay-theme-tasks#gulp-inject#ansi-colors"
```

Just to be safe, I checked the usages of "ansi-colors" in
"plugin-errors" and "gulp-inject" and confirmed that they only use
"safe" colors like "cyan", "magenta" etc (ie. not "bgBlack" etc).