Remove unnecessary parametrization of "normalize()" function
We pass in a `config` object but we only ever access one specific key
of it in hard-coded fashion ("unstyled"), and furthermore, the value
corresponding to the key is always the same.

This is giving an illusion of flexibility that is neither real nor
exercised so let's get rid of the options and just use the string
literal.