Make tag injection idempotent
Working in the monorepo I'm seeing the tags getting injected multiple
times in the "node_modules/liferay-frontend-theme-unstyled" directories.

Rework the "normalize()" function to only inject the tags if they are
not already present.

I've complained elsewhere about the nastiness of mutating installed
node_modules -- ideally we'd change that -- but for now I'm just looking
to apply a bandaid over this unintended side-effect of the current
implementation.