Remove `alternatePath` param from `getPackageJSON()`
There is nowhere in the codebase that we pass a parameter to
`getPackageJSON()`, so we can remove it. We always attempt to find the
package.json relative to the current directory. This means that we can
get rid of the `path.join()` call as well, because `fs.readFileSync()`
is going to operate relative to the current directory in any case.