Don't overwrite divert.defaultVersion from generator
First of all, we can and should use explicit parameter passing (explicit
is better than implicit) here, and it's what we're doing elsewhere in
this file anyway.

Second of all, the method we're wanting to call (`_printWarnings()`)
doesn't even care about the version anyway.

With this change, there is nowhere in the codebase which is reaching in
and mutating a property on the `divert()` function from the outside.

Test plan: `yarn test` and `yo ./packages/generator-liferay-theme`.