Migrate "kickstart_prompt_helpers.js" to use `lookup()`
This is my attempt at replacing the `divert()` pattern with something
slightly less magical and more centralized. Instead of trying to require
a file at runtime whose name we construct dynamically, give everything a
name and look it up by name.

More importantly though, from a maintainability perspective, this
pattern enables us to put related functionality in the same file instead
of spreading it across the filesystem, so it is easier to see the
commonalities and differences between different versions.

For example, in this commit, you can see at a glance that the 7.0
version adds one more choice to the list of choices at the kickstart
prompt, but other than that it is the same.

I'll continue to move the other `divert()`-powered stuff in the
following commits.

Test plan: `yarn test`