Remove dependency on eslint-config-google
As described in the related issue:

> Arguably, the eslint-config-google preset constitutes a great deal
> of somewhat arbitrary fine-tuning that doesn't deliver a convincing
> benefit beyond a much more "stock" or "vanilla" configuration. We
> could streamline things considerably by removing it.

This commit removes the config, and also the commented-out default
values of all the rules; maintaining that list is a fair bit of
administrative overhead that we can do without. The "juice isn't worth
the squeeze", so to speak.

Closed: https://github.com/liferay/eslint-config-liferay/issues/18