Lazily evaluate dependencies in `lookup()`
I didn't do this previously because I wanted to keep things as simple as
I could, and the performance costs of eagerly evaluating all the deps
are negligible because we're talking about so little code. But it came
up in code review as a question:

https://github.com/liferay/liferay-themes-sdk/pull/179#discussion_r261726702

So I figured I may as well add an arrow function here to make the
evaluation lazy. The main property I want to preserve is that the
require targets all be string literals and therefore type-checkable.