Remove aggressive rule deactivations (#25)
These suppressions are a bit aggressive. We're not using them elsewhere,
they override stuff that is on by default in eslint:recommended, and on
the rare occasions when you don't want the linter to complain about a
specific instance, you can add an inline suppression.

Related: https://github.com/liferay/eslint-config-liferay/issues/25