Add .prettierrc.json
I'm about to update the eslint-config-liferay dependency but I've found
that running `npm run format` makes large changes to the source tree,
which makes me think that the absence of a config here is leading us to
fall back to Prettier defaults.

This .prettierrc.json matches the one we're using in a few other places
(eg. alloy-editor, clay, liferay-themes-sdk).