Update eslint, eslint-config-liferay packages
This change ends up requiring many files to be reformatted, so add a
`.prettierrc.json` at the same time, matching the one that we use in
other projects.

I also took the opportunity to change from the deprecated .eslintrc
format to .eslintrc.js.

Some new lints pop up as a result of this change, but I'll fix those in
a separate commit so it remains clear what changes are formatting
changes and which ones are lint fixes.