Adjust formatting scripts in package.json (#279)
Similar to what I did in:

https://github.com/liferay/liferay-amd-loader/pull/192

This commit adds a "format:check" script and uses that in the "ci"
script, to ensure we run the CI checks against the same globs that we're
using locally when running "format"; the globs are adjacent so hopefully
we can keep them in sync (faith springs eternal).

Additionally I removed the unnecessary "packages/*/src/*.js" glob
because the "packages/*/src/**/*.js" glob already matches the same
files, and I added a ".*.js" glob because in the next commit I am going
to migrate the .eslintrc file (which is deprecated) to .eslintrc.js.