Remove "--deprecated" and template language switching (#184)
Remove references to the "--deprecated" switch which may have worked
in the past but definitely doesn't work in current "master" and I am
not sure how long ago it broke. If you look at the code as it existed a
little over a week ago when we started this spike, you can see that the
prompting was not conditional in any way at that point:

https://github.com/liferay/liferay-themes-sdk/blob/d92242148c7a2b2d5f5c56799fea3123804cf743/packages/generator-liferay-theme/generators/app/index.js
https://github.com/liferay/liferay-themes-sdk/blob/d92242148c7a2b2d5f5c56799fea3123804cf743/packages/generator-liferay-theme/lib/divert/common/app_helpers.js

In any case,

the intent was to not show deprecated options like Velocity templates
unless you ran `yo liferay-template --deprecated`, but you can see that
we were offering Velocity templates as an option regardless back then.

As v9 of the toolkit targets DXP 7.2, there is no point in keeping all
this template language stuff around: only Freemarker templates are
supported. So, this diff gets rid of the code that dealt with template
language selection entirely.

Test plan: `yarn test` and create a theme with `yo
./packages/generator-liferay-theme`.

Related: https://github.com/liferay/liferay-themes-sdk/issues/184