Migrate 7.0/upgrade.js to 7.1/upgrade.js (#184)
As per the plan proposed in #184, this commit takes the existing
7.0-to-7.1 upgrade script:

    packages/liferay-theme-tasks/lib/upgrade/7.0/upgrade.js

and moves it to:

    packages/liferay-theme-tasks/lib/upgrade/7.1/upgrade.js

modifying it to become a 7.1-to-7.2 upgrade script. The changes from 7.1
to 7.2 are basically null, which means that most of the script goes
away.

Test plan: `yarn test` — I haven't actually tried upgrading a theme from
the command-line yet because I want to make some additional changes
first (namely, getting rid of the "liferay-theme-deps-7.*" stuff and
replacing it with explicit dependencies).

Related: https://github.com/liferay/liferay-themes-sdk/issues/184