Make getCustomDependencyPath a public method
This is a step towards making "liferay-frontend-common-css" a direct
dependency of "liferay-theme-tasks". At the moment, it is only a
transitive dependency (via "liferay-theme-deps-7.x" packages). We always
depend on the same version, so it may as well be hoisted. Additionally,
in the context of Liferay DXP we end up overriding the paths anyway via
the custom path mechanism, so we will in practice always be using
another version anyway.

Making this method public enables us to require the module directly
where we need it without going through `resolveDependency`, which is
used in other places and it's not easy to change all of those at once.
In the next commit I'll show how I'm going to do that, and then in later
commit(s) I'll tackle `resolveDependency` itself.