"Hoist" liferay-frontend-common-css over to liferay-theme-tasks
As described in the previous commit, we always use the same version of
this module so there is no need for it to be buried in the
"liferay-theme-deps-7.x" packages. And in Liferay DXP, we use the custom
override feature to pass in a specific path in any case.

So, in this commit we:

- Remove "liferay-frontend-common-css" from "liferay-theme-deps-7.x"
  packages.
- Add "liferay-frontend-common-css" to the "liferay-theme-tasks"
  package.
- Update callsites to not use `resolveDependency()` and instead try for
  a custom path before falling back to requiring the
  "liferay-theme-tasks" version.

Note I'm doing it this way because `resolveDependency()` is being used
in other places still and it won't work after "hoisting" the package
(this is because it tries to find the dependency in the theme's own
dependencies, and if it is not there, it tries to find it transitively
in the "liferay-theme-deps-7.x" package; but that won't work any more
because we just removed it from those packages).

For now the two callsites are doing some extra work, but once I've
updated all the `resolveDependency()` callsites (future commits),
then I'll be able to consolidate and simplify things again.