Remove remaining liferay-theme-deps-7.* packages
In the course of working on:

    https://github.com/liferay/liferay-themes-sdk/pull/187

this ends up being a replacement for:

    https://github.com/liferay/liferay-themes-sdk/pull/174

which I previously put up as a draft PR.

Focusing on supporting the latest version only allows us to remove a
fair bit of the dynamism here, and there will be even more that can be
removed in follow-up, but this is a glimpse of a partial step.

Test plan:

`yo ./packages/generator-liferay-theme`; `cd` into theme directory and
update `gulpfile.js` to point at local copy of "liferay-theme-tasks"
(`npm link` also would have done it). Run `gulp build` and see it work
and generated files look right.

There are three failing tests running `yarn test` with this one which I
will follow up and fix, plus some "TODO" comments added for further
stuff to be addressed.