Get deploy tests working again
This is a pretty ugly intermediate step that I am taking as I figure out
what the major test groups need. This fixes 1 of 3 failures (the other
two are in the build and watch tests, and I suspect will require similar
fixes).

Once I've seen all 3 fixes, I'll be in a better place to determine what
the end state should be: I suspect we'll want a thorough review of the
theme fixtures, updating the default options, and refactoring to make
sure the right helpers and in place.