Rework environment variable handling in tests
I noticed that if you "save" an environment variable that is unset (ie.
`undefined`, when you restore it, the variable takes on the value
`"undefined"` (ie. the string, "undefined").

Rather than saving and restoring, then, just set and delete. Custom
paths should be coming in from outside in the test suite anyway.