Fix post-install for layout generator
In testing:

    https://github.com/liferay/liferay-js-themes-toolkit/pull/192

I noticed that our post-install gulp task was bombing
out after creating a new layout. I went back and established that the
same is happening on current "master", due to this change:

    https://github.com/liferay/liferay-js-themes-toolkit/pull/173

Testing the previous release (v8.0.0-rc.3) confirms that this was a
regression.

The way it used to work was that the post-install would run in the
directory of the just-generated layout, and the generated gulpfile in
that directory pulls in "liferay-plugin-node-tasks" and runs "init".

In the buggy "master" branch, however, we were instead running "init"
from "liferay-js-themes-toolkit" (AKA "liferay-themes-sdk"); that is
fine for generated themes, but not for generated theme layouts.

So, the fix here is obvious: run the right gulp task.

If this PR gets delayed for whatever reason I can backport an equivalent
fix onto the "master" branch too.

Just wanted to make some notes here based on some archeology that I did
while exploring this and asking questions in Slack:

- "liferay-theme-tasks" depends on "liferay-plugin-node-tasks".
- Both register a "build" task, so one clobbers the other, I think
  (concretely, I think "liferay-theme-tasks" registers the tasks from
  "liferay-plugin-node-tasks" and then overwrites the latter's "build"
  with its own).
- "liferay-theme-tasks" ends up calling "liferay-plugin-node-tasks" in a
  couple of places; for example, "deploy:gogo" will invoke
  "plugin:deploy-gogo", "deploy:war" will invoke "plugin:deploy" etc.
- Those plugin tasks depend on "build", but due to the overwriting
  mentioned above, I *think* they actually end up calling the theme
  tasks version of "build", which has already run at that point, and at
  least under Gulp v3 (I *think*) won't run again.
- https://www.npmjs.com/package/liferay-plugin-node-tasks claims that
  "liferay-plugin-node-tasks" lives at:
  https://github.com/Robert-Frampton/liferay-plugin-node-tasks
  but it actually lives at:
  https://github.com/jbalsas/liferay-plugin-node-tasks
  (we should correct that link in the next release and ideally move
  this into the "liferay" GitHub org).

Test plan: `yo ./packages/generator-liferay-theme/generators/layout`