Hack around error at end of `gulp init` in generators
Discovered this bug while testing the "import" generator for v8: this
generator imports a 6.2 theme, and when it runs the "init" gulp task at
the end `doctor.js` kicks in and rejects the theme version in the
packag.json because it is looking at argv in order to figure out which
task is running; "some generator name" is not a task, so it hits a
`default` case in a `switch` statement instead of the desired `init`
case, leading an error to be thrown.

The fix here is a tremendous hack, but it will be short-lived because
we're killing the doctor in v9 (see #196).

For reference, this bug was introduced in #173.

Test plan: `yo ./packages/generator-liferay-theme/generators/import` and
provide the absolute path to:

    ./packages/generator-liferay-theme/generators/import/__tests__/__fixtures__/sdk-theme

as a path to import.

Related:

- https://github.com/liferay/liferay-js-themes-toolkit/pull/195
- https://github.com/liferay/liferay-js-themes-toolkit/issues/196