Rename liferay-themes-sdk to liferay-js-themes-toolkit for v9 (#185)
In terms of the documentation and internal terminology we want to
rebrand this as "a themes toolkit" as opposed to an SDK because it is
closer in spirit to a toolkit than a set of APIs, and it's a pattern
we've been following in other places.

Note there was one section of a README that spoke about
liferay-theme-deps packages; these no longer exist in v9, so I
removed/reworked the section rather than just updating the naming from
"sdk" to "toolkit".

The repo itself was already renamed and redirects are in place.

Closes: https://github.com/liferay/liferay-js-themes-toolkit/issues/185

Corresponding v8 change: https://github.com/liferay/liferay-js-themes-toolkit/pull/197