Fix kickstart.test.js breakage
We had an assertion in here that broke due to the addition of the
license header to one of the fixture files in the last commit
(packages/liferay-theme-tasks/test/fixtures/themes/7.0/kickstart-theme/src/js/main.js).

Two possible fixes here:

1. Change the eslint config to not add a header to this file.
2. Change the test assertion to tolerate the presence of the header.

Given that 1 is easier, and we are likely to drop the "kickstart" task
in the future, I took that route.