Remove incomplete DXP 7.2 support from v8 of toolkit
We originally implemented this partially in:

  https://github.com/liferay/liferay-js-themes-toolkit/pull/158

While working on the v9 PR:

  https://github.com/liferay/liferay-js-themes-toolkit/pull/187

I discovered a couple places we had missed, and addressed those in:

  https://github.com/liferay/liferay-js-themes-toolkit/pull/187/commits/4e7e120d0003bbd4cbaf1aa5208c5d4d4a599c29

and:

  https://github.com/liferay/liferay-js-themes-toolkit/pull/187/commits/e3ceaf17649481fec0670cf7a7117b1976ceee06

Since then, we have committed to making a clean split between v8 and v9
in terms of DXP 7.2 support:

  https://github.com/liferay/liferay-js-themes-toolkit/issues/184

So, rather than add the missing 7.2 bits to the v8 series, this commit
removes them. This PR:

  https://github.com/liferay/liferay-js-themes-toolkit/pull/200

adds a v8 README that makes it clear that you need to use v9 in order to
create 7.2 themes. Note that in v9 we removed the "liferay-theme-deps-*"
packages. We never actually published "liferay-theme-deps-7.2" to NPM,
so there is no clean-up required there, other than deleting the package
directory here.

Test plan: `yarn test`, use generator.