Fix "packageVersion is not defined" generator an "All" themelet
This is a regression that I most likely introduced in:

https://github.com/liferay/liferay-js-themes-toolkit/pull/195

The v9 version of the same:

https://github.com/liferay/liferay-js-themes-toolkit/pull/187

doesn't have the same bug because it uses a different workaround, but I
like the one used in this commit better so I am going to propagate it
over there too.

Basically, we can't call `lookup` to get the `devDependencies` when
making an "All" versions themelet, and we don't have to anyway, because
the generated "package.json" does not have a "devDependencies" property
in any case.

Test plan:

- `yarn test`
- `yo ./packages/generator-liferay-theme/generators/themelet`
- `yo ./packages/generator-liferay-theme/generators/import`
- `yo ./packages/generator-liferay-theme/generators/layout`
- `yo ./packages/generator-liferay-theme`