Fold liferay-plugin-node-tasks into liferay-theme-tasks
This is actually step 1: just moving the files in here and getting theme
working. The next step after that will probably be moving them around a
little more.

What I've done here is a straight dump of the files from the upstream
repo:

https://github.com/jbalsas/liferay-plugin-node-tasks

into "packages/liferay-theme-tasks/plugin/". In order to make that
possible I first did:

- https://github.com/wincent/liferay-plugin-node-tasks/pull/2 ("Port
  from ava to jest"); and:
- https://github.com/wincent/liferay-plugin-node-tasks/pull/1 ("Remove
  all references to gogo-shell ")

Then I moved the files in here, removed existing (dead) references to
the "deploy:gogo" and "plugin:deploy-gogo" tasks, added the
"gulp-zip@^3.2.0" dependency (it is the only dependency in
"liferay-plugin-node-tasks" that is not already in
"liferay-theme-tasks"; note that although the latest release it 4.2.0, I
kept the "^3.2.0" version specifier to make this initial change as small
as possible), and updated references that previously pointed at the
plugin package to instead refer to the local version.

Test plan: `yo ./packages/generator-liferay-theme/generators/layout` etc

Note: This commit does not pass our lint checks because I want the
corresponding fixes to be clearly visible in a separate commit; likewise
for formatting fixes.