Apply formatting fixes
With `yarn format` (which, as noted in other commits, ends up being
equivalent to `eslint --fix` because of our use of prettier-eslint, and
so we get some lint fixes in here as well).

Still a few lints to resolve after this one, which aren't autofixable;
I'll keep those in a separate commit.