Update del from v1.2.1 to v4.0.0
This fixes several failures in the plugin tests, taking us from:

    Test Suites: 6 failed, 1 passed, 7 total
    Tests:       1 failed, 17 passed, 18 total

to:

    Test Suites: 2 failed, 5 passed, 7 total
    Tests:       1 failed, 28 passed, 29 total

in `yarn test plugin`.

Prior to the update, tests where failing due to errors being thrown in
the `afterAll` clean-up, where the promise supposedly returned by
`del()` was actually undefined:

    TypeError: Cannot read property 'then' of undefined

Note that in v2 `del` moved to a Promise-based API, so I had to update a
few usage sites elsewhere; calls to `del.sync()` continue to work as
before.

Test plan: `yarn test` and see the 4 fewer failures in the plugin tests
and no regressions in the other tests.