Tweak jest config to properly target plugin tests
Needed to exclude "__snapshots__" to stop Jest from trying to find tests
in a snapshot and complaining like this:

    packages/liferay-theme-tasks/lib/__tests__/__snapshots__/look_and_feel_util.test.js.snap
      ● Test suite failed to run

        Your test suite must contain at least one test.

Also needed to make the "/test/*.js" pattern more specific: we
need this pattern to match the plugin tests, but if not scoped to
"/plugin/test/*.js" it ends up matching test utility files elsewhere
that don't contain tests.

Yes, the config is a mess and we need to make a consistency pass over it
all in the future.