Update run-sequence from v1.0.2/v1.2.2 to v.2.2.1
This was an attempt to fix some test failures based on the fact that the
gulp tasks which were changed a few commits back in the "del" package
upgrade (to return promises) weren't actually being waited for in the
test suite.

Sadly, I must report that the upgrade does not fix this issue, although
we may as well do it anyway. The fault is evidently not that
run-sequence fails to wait for promise-returning tasks: it's that
`gulp.hook('after:*')` runs immediately after the promise-returning task
has returned. if the task uses the callback based async interface
instead, then the hook does run at the right time. Sigh. In the next
commit I'll change those `del`-using tasks to use callbacks instead of
returning promises, and that will fix those issues.

I'd like to report that the vulnerability count is down after this
upgrade, but it remains as before:

    49 vulnerabilities found - Packages audited: 538608
    Severity: 20 Low | 13 Moderate | 16 High