Don't return promises from Gulp tasks
Returning promises should work, but it breaks our tests. As described in
the parent commit, if we return a promise from a task, then the
corresponding `gulp.hook('after:*')` function runs immediately instead
of waiting for the promise to resolve. This breaks our tests because our
assertions end up running too soon.

So, switch back to callback style again, but note that we have to call
the callback from inside a `then()`; the `del()` API changed from
`del([stuff], callback)` to `del([stuff]).then(...)` in v2, and we did a
big update all the way from 1.3.1 to 4.0.0.