Set jest default timeout globally
While working on:

    https://github.com/liferay/liferay-js-themes-toolkit/pull/213

I noticed that the plugin tests would pass in isolation, but fail with a
timeout error if run after running the other tests (due to taking longer
than the default 5s to run). My hunch was that the other tests might be
creating such a large quantity of junk that it was taking more than 5s
to clean it up in a before or after hook.

Hence, this commit, which enables me to delete a bunch of repeated calls
to `jest.setTimeout()`. That's the good news.

The bad news is that it doesn't fix the plugin test failures when not
run in isolation; I'm going to continuing investigating that and will
hopefully have a real fix for it soon.