Remove hideConsole and restoreConsole
The intent of this was obviously to make the test output look "clean",
but in the case of failing tests it only makes debugging harder (for
example, you can't `console.log` intermediate state without first
commenting out the `hideConsole` and `removeConsole` calls), so I'm
going to remove it. Better that it be ugly and debuggable than beautiful
and cryptic.