Don't let promise resolutions be treated as errors
So this is the "final" version of the fix I implemented a few commits
back (in 388399e, "Don't return promises from Gulp tasks"). Some of the
test failures where because `del` would resolve with a list of deleted
files, so when we forwarded that via `cb()`, it was treated as an error.
This is a bit of a foot-gun with this API, but there's not much to be
done about that because we control neither "gulp" nor "del".

Test plan: `yarn test`