Remove liferay-theme-es2015-hook (toolkit v8) (#110, #164, #199)
Three reasons to delete this:

1. It doesn't actually work:
   https://github.com/liferay/liferay-js-themes-toolkit/issues/110
2. We wanted to simplify the structure of the monorepo anyway:
   https://github.com/liferay/liferay-js-themes-toolkit/issues/164
3. It has ancient dependencies which add noise to `yarn audit`:
   https://github.com/liferay/liferay-js-themes-toolkit/issues/199

`yarn audit` goes from:

    105 vulnerabilities found - Packages audited: 139619
    Severity: 38 Low | 31 Moderate | 36 High

to:

    94 vulnerabilities found - Packages audited: 135401
    Severity: 34 Low | 28 Moderate | 32 High

with this change.

We target v8 of the toolbox in this change, but I will make the
equivalent change in the "master" (v9) branch as well.

A later step will be to address #110, but that will involve completely
rethinking and reimplenting how we do ES2015 support.