Remove lodash usage from generator (#124)
Specifically, app/index.js.

Note that I extracted a `normalizeName()` function into a separate
module to replace the existing `deburr` + `kebabCase` functionality,
because we'll need the same thing in one of the sub-generators too. The
approach here is more aggressive than just "deburr"-ing (we strip
anything that is not alphanumeric ASCII), but that is fine because this
is only used to produce a default in a prompt anyway.

Test plan: `yarn test` and `yo ./packages/generator-liferay-theme`