Remove Lerna related stuff
As discussed in the previous commit; we decided to try publishing
without the aid of Lerna and found it to be not too onerous.

What it currently gives us:

- Ability to publish all packages at once.
- Ability to publish only packages that have actually changed since the
  last time.

What it costs us:

- Large dependency footprint.
- Requires us to create and publish additional release branches for
  every version.

Factors to consider for the future:

- The v9 branch only has 3 packages, and only 1 dependency between
  packages, so value of Lerna is diminished there.
- We expect the bulk of future work to be on the v9 branch (that is,
  "master").
- We could still create a tiny "publish" script if we want to retain a
  small degree of automation.

To avoid confusion, I think we should remove the existing references,
and that is what this commit does.

Related: https://github.com/liferay/liferay-js-themes-toolkit/issues/161