Fold lookup/kickstart.js into prompts/kickstart_prompt.js
This module has become simpler as we've trimmed away functionality, to
the point that there is no dynamism involved in it any more (no need to
lookup by version) and therefore no need for it to live under the
`lookup()` system any more.

Test plan: `yarn test`, and, in a generated theme,
`node_modules/.bin/gulp kickstart`. Note that to test the generated
theme you have to update the "package.json" and the "gulpfile.js" to
both reference "../packages/liferay-theme-tasks" (ie. presuming that you
create your test generated theme inside this repo).