Add license headers to JS files (#193)
The license template is in "copyright.js", which "eslint-config-liferay"
automatically picks up and applies (thanks to the "eslint-notice"
plug-in). For simplicity, 2014 is selected as the copyright year because
that's the year of the first commit to the project. In the next commit,
I'll substitute that for a `YEAR` placeholder, which means that new
files created in the repo will automatically use the year of creation
instead.

Closes: https://github.com/liferay/liferay-amd-loader/issues/193