Make `gulp watch` pick up changes in subdirectories (#235)
I suspect the test here has been broken since at least c15b0ed1a20fe. We
were creating a "rootDir" relative to "./src" that might look like
"css" or "css/partials" etc, and then comparing it against "css" instead
of doing a prefix match.

The smallest fix would have been to change those `=== 'css'` style
checks to `match(/^css\b/)` etc, but in the end I went with something
that I hope is clearer and more readable. We add a `getResourceDir()`
function that returns us a string like "css", "js" etc, and we stick
with the `===` comparisons. I went with the "resource" terminology
because I think "root" might be confusing here (root of what? the repo?
the theme? the src directory within the theme?).

Note also that the path operations we are doing here are simpler than
before. Now it's basically:

    relativePath = path.relative('./src', resource);

whereas before it was:

    relativePath = path.relative(
      path.join(process.cwd(), './src'),
      path.dirname(resource)
    );

This is the v8 version of the fix, cherry-picked from:

    https://github.com/liferay/liferay-js-themes-toolkit/pull/237

Test plan: `yarn test` (note that I did manual testing in the v9 change
because I have a "master" instance of portal, but this change applies
cleanly here and I am confident that it will work for v8 too).

Fixes: https://github.com/liferay/liferay-js-themes-toolkit/issues/235