Make `gulp watch` pick up changes in subdirectories (#235)
I suspect the test here has been broken since at least c15b0ed1a20fe. We
were creating a "rootDir" relative to "./src" that might look like
"css" or "css/partials" etc, and then comparing it against "css" instead
of doing a prefix match.

The smallest fix would have been to change those `=== 'css'` style
checks to `match(/^css\b/)` etc, but in the end I went with something
that I hope is clearer and more readable. We add a `getResourceDir()`
function that returns us a string like "css", "js" etc, and we stick
with the `===` comparisons. I went with the "resource" terminology
because I think "root" might be confusing here (root of what? the repo?
the theme? the src directory within the theme?).

Note also that the path operations we are doing here are simpler than
before. Now it's basically:

    relativePath = path.relative('./src', resource);

whereas before it was:

    relativePath = path.relative(
      path.join(process.cwd(), './src'),
      path.dirname(resource)
    );

This is the v9 version of the fix. Will cherry-pick to the 8.x branch as
well.

Test plan: Create a theme with two `@import`-ed SCSS partials, one of
them inside a nested folder. See live-reloading work with `gulp watch`
when either partial is edited.

Fixes: https://github.com/liferay/liferay-js-themes-toolkit/issues/235